[core] Merge uv_detail.hpp into uv.hpp
