+++ platform/default/timer.cpp
@@ -1,7 +1,7 @@
 #include <mbgl/util/timer.hpp>
 
 #include <mbgl/util/run_loop.hpp>
-#include <mbgl/util/uv_detail.hpp>
+#include <mbgl/util/uv.hpp>
 
 namespace mbgl {
 namespace util {
